resolved conflicts

+++ adapter/internal/oasparser/model/api_yaml.go
@@ -116,7 +116,7 @@ func getRateLimitPolicy(throttlingLimit ThrottlingLimit) APIRateLimitPolicy {
 	policyName := GetRLPolicyName(throttlingLimit.RequestCount, throttlingLimit.Unit)
 	var rlType string = "REQUEST_COUNT"
 	rlPolicy.PolicyName = policyName
-	rlPolicy.Count = uint32(throttlingLimit.RequestCount)
+	rlPolicy.Count = throttlingLimit.RequestCount
 	rlPolicy.Type = rlType
 	rlPolicy.Span = 1
 	rlPolicy.SpanUnit = throttlingLimit.Unit
